backend/requirements.txt
# backend/requirements.txt

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6

# OCR libraries
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
# paddleocr==2.7.0
# paddlepaddle==2.5.2

# ML libraries
lightgbm==4.1.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
shap==0.43.0
joblib==1.3.2

# Web3 / Blockchain
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# Storage
boto3==1.29.7
ipfshttpclient==0.8.0a2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
